pyproject.toml
[project]
name = "structured-output"
version = "0.1.0"
description = "Batch processing of legal case rulings using Google Vertex AI and Gemini models"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai",
    "pydantic",
    "python-dotenv",
    "google-cloud-storage",
    "google-cloud-aiplatform",
]

[tool.ruff]
line-length = 100
target-version = "py313"
